Add dn, sin and refactor expiry
This change adds `dn` and `sin` claim support to claims-gen endpoint and
also includes the aggregated values as array for all licenses which have
these claims non-empty in the corresponding product claim sets.
Furthermore the existing expirations claim in the product result set is
renamed to expiry and includes the aggregated expiry claim of all active
licenses for the corresponding product.